* adding animation for adding tiles on top row

+++ core/src/com/match3/game/entities/Tile.java
@@ -13,10 +13,10 @@ public class Tile {
     public float y = 0;
     public boolean isActivated = false;
     public boolean toDestroy = false;
-    public int sizeX = 48; //! tile sizeX in px
-    public int sizeY = 48; //! tile sizeY in px
+    public float sizeX = 48; //! tile sizeX in px
+    public float sizeY = 48; //! tile sizeY in px
 
-    public Tile(TileType type, float x, float y, int sizeX, int sizeY) {
+    public Tile(TileType type, float x, float y, float sizeX, float sizeY) {
         this.type = type;
         this.x = x;
         this.y = y;
